OkRegistrator

Documentation for OkRegistrator.

OkRegistrator.baseexactexpr

OkRegistrator.folderlist

OkRegistrator.folderlist

OkRegistrator.folderlist

OkRegistrator.isprerelease

OkRegistrator.isprerelease

OkRegistrator.okciregister

OkRegistrator.baseexactexpr — Method

baseexactexpr(regrepo) returns expression matching exactly the basename of regrepo.

Example

expr = OkRegistrator.baseexactexpr(".julia/dev/OkRegistry")
occursin(expr, "OkRegistry")

# output

true

occursin(expr, "NotOkRegistry")

# output

false

occursin(expr, "NotOkAtRegistryAll")

# output

false

OkRegistrator.folderlist — Method

folderlist(dir; join=true) return the list of folder but no subfolder under dir. This function uses walkdir.

OkRegistrator.folderlist — Method

folderlist(expr::Regex, dir; join=true) returns a vector of paths who match expr.

OkRegistrator.folderlist — Method

folderlist(;kwargs...) returns results in the current directory (pwd)

OkRegistrator.isprerelease — Method

Given the package directory pkgpath, isprerelease(pkgpath) returns true if there is nothing left after matching r"(\d+.?)+". Use this to check whether to registor a version or not.

OkRegistrator.isprerelease — Method

isprerelease(ver::VersionNumber)

Example

isprerelease(v"0.99.5")

# output

false

isprerelease(v"0.99.5-DEV")

# output

true

OkRegistrator.okciregister — Method

Given the directory of your local registry okregpath, okciregister(okregpath) attempts to register all pacakges (folders) under dirname(okregpath).